backend/requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
autopep8==1.5.6
billiard==3.6.4.0
boto3==1.18.39
botocore==1.21.39
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-braces==1.14.0
django-celery
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-coupons==1.1.0
django-debug-toolbar==3.2.1
django-oauth-toolkit==1.5.0
django-phone-login==1.1.3
django-phonenumber-field==1.1.0
django-redis==4.12.1
django-rest-framework-social-oauth2==1.1.0
django-sendsms==0.2.3
django-sms==0.5.0
django-smtp-ssl==1.0
django-storages==1.11.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
drfpasswordless==1.5.7
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
jwcrypto==0.8
kombu==5.1.0
oauthlib==3.1.0
phonenumberslite==8.12.27
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
sslcommerz-lib==1.0
toml==0.10.2
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
